Add outdatedApi diagnostic, extract migration DB, update README and changeset
