+++ .changeset/outdated-api-diagnostic.md
@@ -1,0 +1,5 @@
+---
+"@effect/language-service": minor
+---
+
+Add `outdatedApi` diagnostic that warns when using outdated Effect APIs in a project targeting a newer version of Effect.
